Split PlayerShip and AIShip to seperate files
If and when the classes are going to get bigger, especially the AI for
NPC ships, separating the classes into their own files should make
Ship.cs much less messy.
